Refactor __future__ code
