+++ maplibre/__future__/datasets.py
@@ -1,9 +1,8 @@
 from typing import Optional, Union
 
 from geopandas import read_file
-from pydantic import BaseModel
-
 from maplibre.sources import GeoJSONSource
+from pydantic import BaseModel
 
 
 class DataSet(BaseModel):
@@ -27,18 +26,3 @@ class DataSets:
         geometry_type="Polygon",
         bounds=(-123.2639151, 49.1995174, -123.0234703, 49.295612),
     )
-
-
-class PaintProperties(BaseModel):
-    line_color: Union[str, list, None]
-    line_opacity: Union[float, int, None]
-    line_width: Union[float, int, None]
-    fill_color: Union[str, list, None]
-    fill_opacity: Union[float, int, None]
-    fill_outline_color: Union[str, list, None]
-    fill_extrusion_color: Union[str, list, None]
-    fill_extrusion_opacity: Union[float, int, None]
-    circle_color: Union[str, list, None]
-    circle_opacity: Union[float, int, None]
-    circle_stroke_color: Union[str, list, None]
-    circle_radius: Union[float, int, None]
